Split composer integration tests from workspace model
