+++ docs/DECISIONS.md
@@ -261,3 +261,4 @@
 - Approval-card presentation keeps raw machine state separate from user-facing copy. Tool cards may remain `status == .review` for automation, tests, and action routing, while `ToolCardReviewState` carries the semantic review substate. Routine approvals display as `Ready` / `Ready to run` with `Run` and `Skip` actions. Only denied or policy-blocked review cards display `Needs review`. Native SwiftUI, static HTML, and the Playwright harness must read this copy from the tool-card surface instead of exposing raw safety jargon.
 - Browser integration coverage lives in `WorkspaceBrowserIntegrationTests`, not `WorkspaceModelTests`. Browser URL normalization, static and fetched HTML snapshots, fetch-failure fallback, comments, history/reload, invalid-address errors, and composer-driven `host.browser.inspect` cross browser engine state, workspace surfaces, tool cards, and transcript events; keep those flows together while pure URL resolution and browser engine reducers stay in focused unit tests.
 - Workspace review integration coverage lives in `WorkspaceReviewIntegrationTests`, not `WorkspaceModelTests`. Apply-patch review diff refresh, local and SSH Remote stage/restore actions, hunk staging, review comments, tool cards, and review-surface visibility cross the workspace model, git tools, remote SSH executor, transcript events, and review surface; keep those flows together while pure review comment planning and git command construction stay in focused unit tests.
+- Composer integration coverage lives in `WorkspaceComposerIntegrationTests`, not `WorkspaceModelTests`. Composer submit, tool-card creation, artifact surfacing, Computer Use dispatch, queued-tool progress, streaming assistant drafts, cancellation notices, empty-draft no-ops, and selection races cross the workspace model, agent runner, transcript events, tool cards, and top-bar status; keep those flows together while pure submission/cancellation planning stays in focused planner tests.
